Extract crypt.c from misc2.c
