+++ src/memline.c
@@ -55,6 +55,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "crypt.h"
 #include "option.h"
 #include "os_unix.h"
 #include "screen.h"
