+++ src/misc2.h
@@ -81,16 +81,6 @@ int illegal_slash(char *name);
 char_u *parse_shape_opt(int what);
 int get_shape_idx(int mouse);
 void update_mouseshape(int shape_idx);
-int crypt_method_from_string(char_u *s);
-int get_crypt_method(buf_T *buf);
-void set_crypt_method(buf_T *buf, int method);
-void crypt_push_state(void);
-void crypt_pop_state(void);
-void crypt_encode(char_u *from, size_t len, char_u *to);
-void crypt_decode(char_u *ptr, long len);
-void crypt_init_keys(char_u *passwd);
-void free_crypt_key(char_u *key);
-char_u *get_crypt_key(int store, int twice);
 int vim_chdir(char_u *new_dir);
 int get_user_name(char_u *buf, int len);
 void sort_strings(char_u **files, int count);
